requirement.txt
bcrypt==4.0.1
cffi==1.15.1
cryptography==38.0.4
future==0.18.2
Jinja2==3.1.2
MarkupSafe==2.1.1
netmiko==4.1.2
ntc-templates==3.1.0
paramiko==2.12.0
pycparser==2.21
PyNaCl==1.5.0
pyserial==3.5
python-dotenv==0.21.0
PyYAML==6.0
scp==0.14.4
six==1.16.0
tenacity==8.1.0
textfsm==1.1.2
